wip(): new proposal
