+++ src/components/Item.tsx
@@ -1,6 +1,6 @@
 import type { FoodObject } from "../types/food";
 
-import { Link } from "react-router-dom";
+import { Link, useParams } from "react-router-dom";
 import { Box, Stack, Typography, styled } from "@mui/material";
 import { useTranslation } from "react-i18next";
 
@@ -11,18 +11,18 @@ const ImgBox = styled(Stack)(({ theme }) => ({
   boxShadow: "3px 4px 8px #888888",
   overflow: "clip",
   color: theme.palette.primary.light,
-  backgroundColor: theme.palette.primary.main, 
-  
+  backgroundColor: theme.palette.primary.main,
 }));
 
 function Item(props: FoodObject) {
   const { t } = useTranslation();
   const image = props.image.toLowerCase();
+  const { selectedMonthNum } = useParams();
   return (
     <Box
       sx={{ flexGrow: 1, width: "100%", maxWidth: "30%", minWidth: 90 }}
     >
-      <Link to={`/foodpage/${props.description[0].slug}`}>
+      <Link to={`/${props.description[0].slug}/${selectedMonthNum}`}>
         <ImgBox>
           <Box position="relative" flexGrow={1}>
             <img
@@ -33,7 +33,6 @@ function Item(props: FoodObject) {
           </Box>
           <Box position="relative" height="20%" sx={{ textAlign: 'center'}}>
             <Typography
-              
               top={0}
               fontSize="max(1rem, min(4.5vw, 3rem))"
               fontWeight={400}
