+++ src/routes/FoodOfTheMonth.tsx
@@ -1,7 +1,7 @@
 import type { FoodList, FoodCategory, FoodObject } from "../types/food";
 import React from "react";
 import { useState, useEffect } from "react";
-import { useParams, Link, useNavigate } from "react-router-dom";
+import { useParams, useNavigate } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { FoodDBContext } from "../contexts/FoodDB";
 import RenderFoods from "../components/RenderFoods";
@@ -9,13 +9,9 @@ import {
   Box,
   Tab,
   Tabs,
-  styled,
-  alpha,
   Chip,
-  Typography,
   Stack,
 } from "@mui/material";
-import { ArrowLeft, ArrowRight } from "@mui/icons-material";
 
 export default function FoodOfTheMonth() {
   const [food, _] = React.useContext(FoodDBContext);
@@ -55,50 +51,13 @@ export default function FoodOfTheMonth() {
     setFoodType(newFoodCategory);
   };
 
-  //variables to change month when pressing the arrows
-  const prevMonth = monthNum != 0 ? monthNum - 1 : 11;
-  const nextMonth = monthNum != 11 ? monthNum + 1 : 0;
-
-  //styled MUI arrows
-  const ArrowButton = styled(Link)(({ theme }) => ({
-    color: alpha(theme.palette.primary.light, 0.75),
-    "&:hover": {
-      color: alpha(theme.palette.primary.light, 0.95),
-    },
-    display: "flex",
-    alignItems: "center",
-    margin: theme.spacing(1),
-    width: "auto",
-  }));
-
-  const ShadowBox = styled(Stack)(({ theme }) => ({
-    boxShadow: `0 2px 4px ${theme.palette.secondary}`,
-  }));
-
   return (
     <Stack height="100%">
-      <ShadowBox
-        direction="row"
-        justifyContent="space-between"
-        bgcolor="secondary.main"
-        color="primary.light"
-        boxShadow="0 2px 4px #332323"
-      >
-        <ArrowButton to={`/month/${prevMonth + 1}`}>
-          <ArrowLeft />
-        </ArrowButton>
-        <Typography variant="h6" display="flex" alignItems="center">
-          {t(`month_${monthNum}`)}
-        </Typography>
-        <ArrowButton to={`/month/${nextMonth + 1}`}>
-          <ArrowRight />
-        </ArrowButton>
-      </ShadowBox>
       <Tabs
         value={foodType}
         onChange={(_, value) => handleChange(value)}
         variant="fullWidth"
-        sx={{ fontWeight: 700, mt: 1 }}
+        sx={{ fontWeight: 700 }}
         aria-label="tabs for the selection of fruits, vegetables or others"
       >
         <Tab
